+++ app/rss/blog/feed.json/route.ts
@@ -1,0 +1,11 @@
+import { generateRssFeed } from "../generate";
+
+export const dynamic = "error";
+
+export async function GET() {
+    return new Response((await generateRssFeed()).json1(), {
+        headers: {
+            "Content-Type": "application/json",
+        },
+    });
+}